+++ package.json
@@ -5,6 +5,10 @@
     "scripts": {},
     "private": true,
     "dependencies": {
+        "@emotion/react": "11.11.1",
+        "@emotion/styled": "11.11.0",
+        "@mui/icons-material": "5.14.13",
+        "@mui/material": "5.14.13",
         "@nx/next": "16.10.0",
         "next": "13.4.1",
         "react": "18.2.0",
