--- portal/app/page.module.css
@@ -1,1 +1,0 @@
-.page {}
